Strip extra Parsoid response if present
For no reason I can discern, on the testing instance in Labs Parsoid is
not respecting the body=1 post parameter and returning a full html
document instead of the requested body fragment. Rather than spend a lot
of time debugging this issue I'm going to cheat and adjust the regex
used to strip the body tag so that it will also strip any additional
surrounding document.

Change-Id: If5c68095fea70778cd0b6e33e83a001aeaeacbc1