Add "composer test" command to lint, check style and run tests
You need to fetch the development dependencies to run tests, by running
`composer update --dev` first. Doing this will dirty the local
composer.lock and vendor directories because we have the `--no-dev`
version of these files checked into the repository. This is a bit of
a mess for the developer to deal with. A possible long term fix would be
to create and maintain a separate repo of the vendor files for
production use.

Change-Id: Ie8db3e9e1e4ddf9e0a1c30f0b12095388a7dd51a